+++ core/src/com/unciv/logic/automation/city/ConstructionAutomation.kt
@@ -98,7 +98,8 @@ class ConstructionAutomation(val cityConstructions: CityConstructions) {
             cityConstructions.getRemainingWork(choice), cityConstructions.productionForConstruction(choice)))
     }
 
-    private fun Sequence<INonPerpetualConstruction>.filterBuildable(): Sequence<INonPerpetualConstruction> {
+
+    private fun <T:INonPerpetualConstruction> Sequence<T>.filterBuildable(): Sequence<T> {
         return this.filter {
             val cache = if (it is Building) buildableBuildings else buildableUnits
             if (cache[it.name] == null) {
@@ -198,7 +199,7 @@ class ConstructionAutomation(val cityConstructions: CityConstructions) {
 
         // Is there already a Workboat nearby?
         // todo Still ignores whether that boat can reach the not-yet-found tile to improve
-        val twoTurnsMovement = buildableWorkboatUnits.maxOf { (it as BaseUnit).movement } * 2
+        val twoTurnsMovement = buildableWorkboatUnits.maxOf { it.movement } * 2
         fun MapUnit.isOurWorkBoat() = cache.hasUniqueToCreateWaterImprovements && this.civ == this@ConstructionAutomation.civInfo
         val alreadyHasWorkBoat = city.getCenterTile().getTilesInDistanceRange(1..twoTurnsMovement)
             .any { it.civilianUnit?.isOurWorkBoat() == true }
@@ -260,7 +261,7 @@ class ConstructionAutomation(val cityConstructions: CityConstructions) {
     }
 
     private fun addBuildingChoices() {
-        for (building in buildings.filterBuildable() as Sequence<Building>) {
+        for (building in buildings.filterBuildable()) {
             if (building.isWonder && city.isPuppet) continue
             addChoice(relativeCostEffectiveness, building.name, getValueOfBuilding(building))
         }